+++ src/compressed_tensors/utils/__init__.py
@@ -13,4 +13,7 @@
 # limitations under the License.
 # flake8: noqa
 
+from .helpers import *
+from .permutations_24 import *
 from .safetensors_load import *
+from .semi_structured_conversions import *
